requirements.txt
click==6.6
Flask==0.11.1
Flask-Redis==0.3.0
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.3.1
PyYAML==3.11
redis==2.10.5
Werkzeug==0.11.10
